moved mappings into separate file
