+++ lib/data-import/definition.rb
@@ -1,4 +1,5 @@
 require 'data-import/definition/lookup'
+require 'data-import/definition/mappings'
 require 'data-import/definition/simple'
 
 module DataImport
